Replaces size_t with uint32_t do get improve compatibility with other modules

+++ vitis/software/Baremetal/src/uz/uz_matrix/uz_matrix.c
@@ -20,7 +20,7 @@
 #include <stdbool.h>
 #include <string.h> // for memcpy
 
-uz_matrix_t *uz_matrix_init(uz_matrix_t *self, float *data, size_t length_of_data, size_t rows, size_t columns)
+uz_matrix_t *uz_matrix_init(uz_matrix_t *self, float *data, uint32_t length_of_data, uint32_t rows, uint32_t columns)
 {
     uz_assert_not_NULL(self);
     uz_assert_false(self->length_of_data);
@@ -34,14 +34,14 @@ uz_matrix_t *uz_matrix_init(uz_matrix_t *self, float *data, size_t length_of_dat
     return (self);
 }
 
-size_t uz_matrix_get_number_of_rows(uz_matrix_t const *const self)
+uint32_t uz_matrix_get_number_of_rows(uz_matrix_t const *const self)
 {
     uz_assert_not_NULL(self);
     uz_assert(self->length_of_data);
     return (self->rows);
 }
 
-size_t uz_matrix_get_number_of_columns(uz_matrix_t const *const self)
+uint32_t uz_matrix_get_number_of_columns(uz_matrix_t const *const self)
 {
     uz_assert_not_NULL(self);
     uz_assert(self->length_of_data);
@@ -61,16 +61,16 @@ void uz_matrix_elementwise_product(uz_matrix_t const *const A, uz_matrix_t const
     uz_assert(A->rows == B->rows);
     uz_assert(A->rows == C_out->rows);
 
-    for (size_t row = 0; row < A->rows; row++)
+    for (uint32_t row = 0; row < A->rows; row++)
     {
-        for (size_t column = 0; column < A->columns; column++)
+        for (uint32_t column = 0; column < A->columns; column++)
         {
             C_out->data[(row * A->columns) + column] = A->data[(row * A->columns) + column] * B->data[(row * A->columns) + column];
         }
     }
 }
 
-float uz_matrix_get_element_zero_based(uz_matrix_t const *const A, size_t row, size_t column)
+float uz_matrix_get_element_zero_based(uz_matrix_t const *const A, uint32_t row, uint32_t column)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
@@ -79,7 +79,7 @@ float uz_matrix_get_element_zero_based(uz_matrix_t const *const A, size_t row, s
     return (A->data[(row * A->columns) + column]);
 }
 
-void uz_matrix_set_element_zero_based(uz_matrix_t *const A, float x, size_t row, size_t column)
+void uz_matrix_set_element_zero_based(uz_matrix_t *const A, float x, uint32_t row, uint32_t column)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
@@ -101,9 +101,9 @@ void uz_matrix_sum(uz_matrix_t const *const A, uz_matrix_t const *const B, uz_ma
     uz_assert(A->rows == B->rows);
     uz_assert(A->rows == C_out->rows);
 
-    for (size_t row = 0; row < A->rows; row++)
+    for (uint32_t row = 0; row < A->rows; row++)
     {
-        for (size_t column = 0; column < A->columns; column++)
+        for (uint32_t column = 0; column < A->columns; column++)
         {
             C_out->data[(row * A->columns) + column] = A->data[(row * A->columns) + column] + B->data[(row * A->columns) + column];
         }
@@ -120,9 +120,9 @@ float uz_matrix_dot_product(uz_matrix_t const *const A, uz_matrix_t const *const
     uz_assert(A->rows == 1U);
     uz_assert(B->rows == 1U);
     float C = 0.0f;
-    for (size_t row = 0; row < A->rows; row++)
+    for (uint32_t row = 0; row < A->rows; row++)
     {
-        for (size_t column = 0; column < A->columns; column++)
+        for (uint32_t column = 0; column < A->columns; column++)
         {
             C += A->data[(row * A->columns) + column] * B->data[(row * A->columns) + column];
         }
@@ -149,14 +149,14 @@ void uz_matrix_multiply(uz_matrix_t const *const A, uz_matrix_t const *const B, 
     // The following implementation is "slow" as in it does not use special mechanism to speed it up. See the following resources for possible improvements.
     // https://github.com/deuxbot/fast-matrix-multiplication/blob/master/mxm.c
     // https://en.wikipedia.org/wiki/Matrix_multiplication
-    size_t m = A->rows;
-    size_t n = A->columns;
-    size_t p = B->columns;
-    for (size_t i = 0; i < m; i++)
+    uint32_t m = A->rows;
+    uint32_t n = A->columns;
+    uint32_t p = B->columns;
+    for (uint32_t i = 0; i < m; i++)
     {
-        for (size_t j = 0; j < p; j++)
+        for (uint32_t j = 0; j < p; j++)
         {
-            for (size_t k = 0; k < n; k++)
+            for (uint32_t k = 0; k < n; k++)
             {
                 C_out->data[(p * i) + j] += A->data[(n * i) + k] * B->data[(p * k) + j];
             }
@@ -168,9 +168,9 @@ void uz_matrix_set_zero(uz_matrix_t *const A)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
-    for (size_t row = 0; row < A->rows; row++)
+    for (uint32_t row = 0; row < A->rows; row++)
     {
-        for (size_t column = 0; column < A->columns; column++)
+        for (uint32_t column = 0; column < A->columns; column++)
         {
             A->data[(row * A->columns) + column] = 0.0f;
         }
@@ -185,9 +185,9 @@ void uz_matrix_add(uz_matrix_t const *const A, uz_matrix_t *const C_out)
     uz_assert(C_out->length_of_data);
     uz_assert(A->columns == C_out->columns);
     uz_assert(A->rows == C_out->rows);
-    for (size_t row = 0; row < A->rows; row++)
+    for (uint32_t row = 0; row < A->rows; row++)
     {
-        for (size_t column = 0; column < A->columns; column++)
+        for (uint32_t column = 0; column < A->columns; column++)
         {
             C_out->data[(row * A->columns) + column] += A->data[(row * A->columns) + column];
         }
@@ -198,7 +198,7 @@ void uz_matrix_add_scalar(uz_matrix_t *const A, float scalar)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
-    for (size_t i = 0; i < (A->rows * A->columns); i++)
+    for (uint32_t i = 0; i < (A->rows * A->columns); i++)
     {
         A->data[i] += scalar;
     }
@@ -208,7 +208,7 @@ void uz_matrix_multiply_by_scalar(uz_matrix_t *const A, float scalar)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
-    for (size_t i = 0; i < (A->rows * A->columns); i++)
+    for (uint32_t i = 0; i < (A->rows * A->columns); i++)
     {
         A->data[i] = A->data[i] * scalar;
     }
@@ -219,7 +219,7 @@ void uz_matrix_apply_function_to_each_element(uz_matrix_t *const A, float (*f)(f
     uz_assert_not_NULL(A);
     uz_assert_not_NULL(f);
     uz_assert(A->length_of_data);
-    for (size_t i = 0; i < (A->rows * A->columns); i++)
+    for (uint32_t i = 0; i < (A->rows * A->columns); i++)
     {
         A->data[i] = f(A->data[i]);
     }
@@ -229,9 +229,9 @@ float uz_matrix_get_max_value(uz_matrix_t const *const A)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
-    size_t length_of_data = A->columns * A->rows;
+    uint32_t length_of_data = A->columns * A->rows;
     float maximum_value = A->data[0U]; // Set maximum arbitrarily to first element to initialize the variable
-    for (size_t i = 1U; i < length_of_data; i++)
+    for (uint32_t i = 1U; i < length_of_data; i++)
     { // Already got the value of the first element [0], thus start at 1
         if (A->data[i] > maximum_value)
         {
@@ -241,14 +241,14 @@ float uz_matrix_get_max_value(uz_matrix_t const *const A)
     return maximum_value;
 }
 
-size_t uz_matrix_get_max_index(uz_matrix_t const *const A)
+uint32_t uz_matrix_get_max_index(uz_matrix_t const *const A)
 {
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
-    size_t length_of_data = A->columns * A->rows;
+    uint32_t length_of_data = A->columns * A->rows;
     float maximum_value = A->data[0U]; // Set maximum arbitrarily to first element to initialize the variable
-    size_t max_index = 0U;
-    for (size_t i = 1U; i < length_of_data; i++)
+    uint32_t max_index = 0U;
+    for (uint32_t i = 1U; i < length_of_data; i++)
     { // Already got the value of the first element [0], thus start at 1
         if (A->data[i] > maximum_value)
         {
@@ -264,17 +264,17 @@ void uz_matrix_transpose(uz_matrix_t *A)
     uz_assert_not_NULL(A);
     uz_assert(A->length_of_data);
     uz_assert((A->rows * A->columns) > 0U); // Guard for VLA
-    size_t old_rows = A->rows;
-    size_t old_columns = A->columns;
+    uint32_t old_rows = A->rows;
+    uint32_t old_columns = A->columns;
     // cppcheck-suppress misra-c2012-18.8 // VLA are not allowed, but are here the better alternative to using malloc. In-place matrix transposition for non-square matrix is an better, not implemented alternative
     float helper_data[A->rows * A->columns]; // VLA to store
     void *ptr_res = memcpy(helper_data, A->data, (A->rows * A->columns) * sizeof(float));
     uz_assert_not_NULL(ptr_res); // checks return value of memcpy to make sure something happend
     A->rows = old_columns;
     A->columns = old_rows;
-    for (size_t row = 0U; row < old_rows; row++)
+    for (uint32_t row = 0U; row < old_rows; row++)
     {
-        for (size_t column = 0U; column < old_columns; column++)
+        for (uint32_t column = 0U; column < old_columns; column++)
         {
             float x = helper_data[(row * old_columns) + column];
             uz_matrix_set_element_zero_based(A, x, column, row);
