+++ vitis/software/Baremetal/test/uz/uz_matrix/test_uz_matrix_dot_product.c
@@ -14,8 +14,8 @@ void tearDown(void)
 }
 
 void test_uz_matrix_dot_product_two_column_vectors(void){
-    size_t rows=1;
-    size_t columns=4;
+    uint32_t rows=1;
+    uint32_t columns=4;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     uz_matrix_t* A=init_array_test_helper(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
@@ -25,8 +25,8 @@ void test_uz_matrix_dot_product_two_column_vectors(void){
 }
 
 void test_uz_matrix_dot_product_two_row_vectors(void){
-    size_t rows=4;
-    size_t columns=1;
+    uint32_t rows=4;
+    uint32_t columns=1;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     uz_matrix_t* A=init_array_test_helper(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
@@ -35,8 +35,8 @@ void test_uz_matrix_dot_product_two_row_vectors(void){
 }
 
 void test_uz_matrix_dot_product_wrong_dimensions(void){
-    size_t rows=2;
-    size_t columns=2;
+    uint32_t rows=2;
+    uint32_t columns=2;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     uz_matrix_t* A=init_array_test_helper(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
