+++ vitis/software/Baremetal/test/uz/uz_nn/test_uz_nn_activation_functions.c
@@ -34,7 +34,7 @@ void test_uz_nn_activation_functions_relu(void)
 {
     float x[5] = {1.1f, -1.2f, -5.0f, 10.0f, 10.3f};
     float expected[5] = {1.1f, 0.0f, 0.0f, 10.0f, 10.3f};
-    for (size_t i = 0U; i < 5U; i++)
+    for (uint32_t i = 0U; i < 5U; i++)
     {
         float y = uz_nn_activation_function_relu(x[i]);
         TEST_ASSERT_EQUAL_FLOAT(expected[i], y);
@@ -45,7 +45,7 @@ void test_uz_nn_activation_functions_relu_derivative(void)
 {
     float x[5] = {1.1f, -1.2f, -5.0f, 10.0f, 10.3f};
     float expected[5] = {1.0f, 0.0f, 0.0f, 1.0f, 1.0f};
-    for (size_t i = 0U; i < 5U; i++)
+    for (uint32_t i = 0U; i < 5U; i++)
     {
         float y = uz_nn_activation_function_relu_derivative(x[i]);
         TEST_ASSERT_EQUAL_FLOAT(expected[i], y);
@@ -57,7 +57,7 @@ void test_uz_nn_activation_functions_leaky_relu(void)
     float x[5] = {1.1f, -1.2f, -5.0f, 10.0f, 10.3f};
     float expected[5] = {1.1f, 0.01f * (-1.2f), 0.01f * (-5.0f), 10.0f, 10.3f};
     float alpha = 0.01f;
-    for (size_t i = 0U; i < 5U; i++)
+    for (uint32_t i = 0U; i < 5U; i++)
     {
         float y = uz_nn_activation_function_leaky_relu(x[i], alpha);
         TEST_ASSERT_EQUAL_FLOAT(expected[i], y);
@@ -69,7 +69,7 @@ void test_uz_nn_activation_functions_leaky_relu_derivative(void)
     float x[5] = {1.1f, -1.2f, -5.0f, 10.0f, 10.3f};
     float alpha = 0.01f;
     float expected[5] = {1.0f, 0.01f, 0.01f, 1.0f, 1.0f};
-    for (size_t i = 0U; i < 5U; i++)
+    for (uint32_t i = 0U; i < 5U; i++)
     {
         float y = uz_nn_activation_function_leaky_relu_derivative(x[i], alpha);
         TEST_ASSERT_EQUAL_FLOAT(expected[i], y);
